% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_OA.R
\name{assign_oa}
\alias{assign_oa}
\title{Assign Output Areas}
\usage{
assign_oa(injuries, oa)
}
\arguments{
\item{injuries}{sf injury dataset}

\item{oa}{sf Output Area boundaries}
}
\value{
sf object with OA codes
}
\description{
Spatially assigns injuries to Output Areas.
}
