% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_injuries.R
\name{create_injuries}
\alias{create_injuries}
\title{Create Injury-Level Dataset}
\usage{
create_injuries(collisions, casualties, vehicles)
}
\arguments{
\item{collisions}{Data frame of STATS19 collisions}

\item{casualties}{Data frame of STATS19 casualties}

\item{vehicles}{Data frame of STATS19 vehicles}
}
\value{
sf object of injury-level dataset
}
\description{
Joins collisions, casualties and vehicles and creates a unique injury_id.
}
