% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_roads.R
\name{match_roads}
\alias{match_roads}
\title{Match Injuries to Road Network}
\usage{
match_roads(injuries, roads, threshold_same = 50, threshold_fallback = 100)
}
\arguments{
\item{injuries}{sf object of injury points}

\item{roads}{sf object of OS Open Roads}

\item{threshold_same}{Numeric distance threshold (default 50m)}

\item{threshold_fallback}{Numeric fallback threshold (default 100m)}
}
\value{
sf object with matched road attributes
}
\description{
Spatially matches injury points to OS Open Roads segments.
}
